Creating our context
Created a context with useContext() hook to interact with our piece of stateand to reduce the amount of props we're passing down. However, there's a performance issue: our components is rerendering each time our state updates. The solution is to implement a separate context for our state and the methods we're using to update it.